refactor: split ShellExecutionService and SessionService into focused modules
Extract types, cleanup strategies, and conversation utilities into
dedicated files to improve maintainability and separation of concerns:
- shellTypes.ts: Shell execution interfaces (ShellExecutionResult, etc.)
- processCleanup.ts: Cross-platform process cleanup strategies
- conversationUtils.ts: API history building and telemetry replay
